shell判断目录不能带'', 将./bash中的sh脚本ln到~/.local/bin

+++ install.sh
@@ -7,21 +7,35 @@ bash dconf/load_dconf.sh terminal "/org/gnome/terminal/legacy/" "$home_path/dcon
 # 执行默认dconf配置
 bash dconf/dconf.sh
 
-rm ~/.vimrc
-ln -s ${home_path}/vimrc ~/.vimrc
+rm ${HOME}/.vimrc
+ln -s ${home_path}/vimrc ${HOME}/.vimrc
 
 # w3m浏览器配置
-if [[ ! -d "~/.w3m" ]]; then
-   mkdir ~/.w3m
+if [ ! -d ${HOME}/.w3m ]; then
+    mkdir ${HOME}/.w3m
 fi
 
-rm ~/.w3m/keymap
-rm ~/.w3m/config
-ln -s ${home_path}/config/w3m-keymap ~/.w3m/keymap
-ln -s ${home_path}/config/w3m-config ~/.w3m/config
+rm ${HOME}/.w3m/keymap
+rm ${HOME}/.w3m/config
+ln -s ${home_path}/config/w3m-keymap ${HOME}/.w3m/keymap
+ln -s ${home_path}/config/w3m-config ${HOME}/.w3m/config
+
+# 将./bash 下的文件软连接到 ~/.local/bin
+if [ ! -d ${HOME}/.local/bin ]; then
+    mkdir -p ${HOME}/.local/bin
+fi
+for sh_file_path in $home_path/bash/* ; do
+    target_ln_path="${HOME}/.local/bin/${sh_file_path##*/}"
+    if [ -f $target_ln_path ]; then
+        rm $target_ln_path
+    fi
+    ln -s $sh_file_path $target_ln_path
+done
 
 # python软连接:coc-jedi使用
-sudo ln -s /usr/bin/python3 /usr/bin/python
+if [ ! -f /usr/bin/python ]; then
+    sudo ln -s /usr/bin/python3 /usr/bin/python
+fi
 # w3m命令行查看图片
 # sudo apt install -y  w3m-inline-image
 sudo apt install -y w3m w3m-img
